Rename `protocol::object_safe` to `boxed_round`
